chore(firebase): move firebase context to single file convention
