--- package/src/firebase/firebase-app-context.ts
@@ -1,10 +1,0 @@
-import { FirebaseApp } from 'firebase/app';
-import { createContext } from 'react';
-
-type FirebaseAppContextValue = {
-  app?: FirebaseApp;
-};
-
-export const FirebaseAppContext = createContext<FirebaseAppContextValue>({
-  app: undefined,
-});
