+++ package/src/firebase/firebase-app-context.tsx
@@ -1,6 +1,15 @@
-import { PropsWithChildren } from 'react';
-import { FirebaseAppContext } from './firebase-app-context';
 import { FirebaseApp } from 'firebase/app';
+import { createContext, PropsWithChildren, useContext } from 'react';
+
+type FirebaseAppContextValue = {
+  app?: FirebaseApp;
+};
+
+export const FirebaseAppContext = createContext<FirebaseAppContextValue>({
+  app: undefined,
+});
+
+export const useFirebaseAppContext = () => useContext(FirebaseAppContext);
 
 type FirebaseAppProviderProps = PropsWithChildren<{
   app: FirebaseApp;
