--- package/src/firebase/use-firebase-app-context.ts
@@ -1,4 +1,0 @@
-import { useContext } from 'react';
-import { FirebaseAppContext } from './firebase-app-context';
-
-export const useFirebaseAppContext = () => useContext(FirebaseAppContext);
